2.3

+++ 2.3.py
@@ -1,5 +1,5 @@
-base = int(input("Anna kanta: "))
-height = int(input("Anna korkeus: "))
+base = float(input("Anna kanta: "))
+height = float(input("Anna korkeus: "))
 
 perimeter = 2 * base + 2 * height
 area = base * height
